return false if google results empty
Sometimes google translate would not be able to translate a string so
it would return an empty string and this functions would in that case
try to do a ‘foreach’ on the string, which would usually, in PHP, be
returned as a fatal error “Cannot access empty property” or so. In that
case, maybe it is better to avoid it and return false.